% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarefullyCausal.R
\name{CarefullyCausal}
\alias{CarefullyCausal}
\title{CarefullyCausal}
\usage{
CarefullyCausal(
  formula,
  data,
  family = "gaussian",
  exposure,
  pvalue = FALSE,
  interaction = NULL,
  boot1 = 100,
  boot2 = 100,
  standardization = TRUE,
  bins = 10,
  confidence = "norm",
  result_type = "log",
  outcome_SL_library = c("SL.glm"),
  ps_SL_library = NULL,
  ps_method = "SL",
  ps_formula = NULL,
  ps_tmle = NULL,
  ip_weights_iptw = NULL
)
}
\arguments{
\item{formula}{Specify your formula as in (Y ~ X + W)}

\item{data}{input your data set}

\item{family}{Can be "gaussian" or "binomial" and relates to the outcome of interest (dependent variable)}

\item{exposure}{Specify your exposure variable name (e.g. "treatment")}

\item{pvalue}{TRUE or FALSE, by default is FALSE and does not show the p-value}

\item{interaction}{Specify the interaction you want to include betwen the exposure and covariate. This is an argument for S-standardization.}

\item{boot1}{The number of bootstrap iterations to get the standard error for S-and-T-standardization}

\item{boot2}{The number of bootstrap iterations to get the p-value for S-and-T-standardization}

\item{standardization}{TRUE or FALSE, to show the S-and-T-standardization. Computationally more intensive due to bootstrap.}

\item{bins}{numeric, relevant when exposure is continuous. The number of bins your exposure will be categorized into.}

\item{confidence}{Can be "norm" or "bca", which correspond to a normal confidence interval or bias-adjusted and corrected confidence interval.}

\item{result_type}{Can be "log" (default), "rr" (risk ratio) or "or" (odd ratio)}

\item{outcome_SL_library}{Specify what estimators to use within Superlearner to estimate the outcome (g-estimation step within TMLE)}

\item{ps_SL_library}{Specify what estimators to use within Superlearner to estimate the propensity scores within TMLE}

\item{ps_method}{Can be "SL" (default) or "cbps" if ps_tmle is NULL.}

\item{ps_formula}{Specify the formula to estimate the propensity scores within TMLE. If empty, is uses same formula that is specified under argument formula with dependent variable being exposure.}

\item{ps_tmle}{It is NULL by default, user can input own estimated propensity scores. It is an n x 1 vector.}

\item{ip_weights_iptw}{user-specified IP weights that are used in IPTW.}
}
\value{
Prints the estimand, estimates, assumptions and diagnostics
}
\description{
Performs outcome regression, IPTW, S-standardization, T-standardization and TMLE.
It provides the estimand, estimates, discusses the assumptions and provides diagnostics including interpretations.
It can be used when having a fixed-exposure, outcome is either dichotomous or continuous and when
exposure is dichotomous, multi-value or continuous.
}
