% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_covb_plot.R
\name{gps_plot}
\alias{gps_plot}
\title{Covariate balance plots}
\usage{
gps_plot(h)
}
\arguments{
\item{h}{Input a CarefullyCausal object}
}
\value{
Covariate balance plots
}
\description{
It checks covariate balance after running CarefullyCausal when exposure is continuous
}
