% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_continuous.R
\name{print.cccont}
\alias{print.cccont}
\title{Output print function (continuous exposure)}
\usage{
\method{print}{cccont}(h)
}
\arguments{
\item{h}{CarefullyCausal object}
}
\value{
Prints all output of Carefully Causal object
}
\description{
Output print function (continuous exposure)
}
