% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plot.R
\name{ps.plot}
\alias{ps.plot}
\title{Propensity Score Density Plots}
\usage{
ps.plot(h)
}
\arguments{
\item{h}{Input a CarefullyCausal object}
}
\value{
Propensity score density plots per contrasts with respect to the reference exposure level
}
\description{
Provides propensity score density plots with respect to the reference exposure category. Applicable when exposure is discrete.
}
