% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_table.R
\name{ps.table}
\alias{ps.table}
\title{Propensity Scores Range Table}
\usage{
ps.table(h)
}
\arguments{
\item{h}{Input a CarefullyCausal object}
}
\value{
Propensity Scores minimum and maximum values per exposure level
}
\description{
Only applicable for discrete exposure variable
}
