% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_plot.R
\name{ridge.plot}
\alias{ridge.plot}
\title{Ridge plots}
\usage{
ridge.plot(h, bins = 10)
}
\arguments{
\item{h}{Input a CarefullyCausal object}

\item{bins}{Number of categories to split the exposure variable}
}
\value{
Ridge plot, conditional density propensity score plots
}
\description{
Applicable for continuous exposure. It first bins the exposure into n (bins argument) categories and plots its conditional density propensity score plots
}
