% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarefullyCausal.R
\name{CarefullyCausal}
\alias{CarefullyCausal}
\title{CarefullyCausal}
\usage{
CarefullyCausal(
  formula,
  data,
  family = "gaussian",
  exposure,
  pvalue = FALSE,
  interaction = NULL,
  boot1 = 100,
  boot2 = 100,
  standardization = TRUE,
  bins = 10,
  confidence = "norm",
  result_type = "log",
  outcome_SL_library = c("SL.glm"),
  ps_SL_library = NULL,
  ps_method = "SL",
  ps_formula = NULL,
  ps_tmle = NULL,
  ip_weights_iptw = NULL
)
}
\arguments{
\item{formula}{specify your formula as in \code{Y ~ X + W} with \verb{Y=outcome of interest (numeric vector)}, \verb{X=exposure (factor or continuous)} and \verb{W=set of variables to adjust for}}

\item{data}{input your data frame}

\item{family}{a description of the error distribution and link function. It expects a character string, either \code{"gaussian"} or \code{"binomial"} depending on specified outcome of interest}

\item{exposure}{specify the exposure variable as character string (e.g.\code{"treatment"})}

\item{pvalue}{\code{TRUE} or \code{FALSE}, by default is \code{FALSE} and does not show the p-value}

\item{interaction}{Interactions that are included specifically in the S-standardization method. These are interactions
between covariates and the exposure. General interaction terms or effect modifications should be specified in the \code{formula} argument}

\item{boot1}{the number of bootstrap iterations to get the standard error for S-standardization and T-standardization. By default set to 100 iterations}

\item{boot2}{the number of bootstrap iterations to get the p-value for S-standardization and T-standardization. By default set to 100 iterations}

\item{standardization}{\code{TRUE} or \code{FALSE}, to show S-standardization and T-standardization estimators. Default is \code{TRUE}, but computationally intensive due to bootstrap and thus can be set to \code{FALSE}}

\item{bins}{\code{integer}, only relevant when exposure is continuous. Specify number of bins your exposure will be categorized into for Ridgeline plot. Default is 10 bins}

\item{confidence}{can be \code{norm} or \code{bca}, which correspond to a normal confidence interval or bias-corrected and accelerated confidence interval. Default is \code{norm}}

\item{result_type}{estimates table can be shown in terms of \code{"log"} (log odds-ratio, default), \code{"rr"} (risk ratio) or \code{"or"} (odds ratio)}

\item{outcome_SL_library}{Specify what estimators to use within Superlearner, when using SuperLearner for g-estimation step in TMLE. Default is \code{SL.glm}}

\item{ps_SL_library}{Specify what estimators to use within Superlearner to estimate the propensity scores within TMLE. Default is same library as specified in \code{outcome_SL_library}}

\item{ps_method}{Specify what method to use for propensity score estimation in TMLE. Expects a character string \code{"SL"} (default) or \code{"cbps"} if \code{ps_tmle=NULL}. When user inputs own propensity scores, \code{ps_tmle!=NULL}, this argument is ignored}

\item{ps_formula}{Specify the formula to estimate the propensity scores in TMLE. If \code{NULL}, it uses same formula that is specified under argument \code{formula} with the \code{exposure} variable being the outcome variable}

\item{ps_tmle}{Input own estimated propensity scores. Expects a vector with the propensity scores. By default this is \code{NULL} and are estimated using method specified at \code{ps_method}}

\item{ip_weights_iptw}{Input own estimated IP weights. Expects a vector containing the IP weights. By default this is \code{NULL} and are estimated using \code{CBPS}}
}
\value{
It prints the causal estimand, causal estimates (output table), causal assumptions and relevant diagnostics
}
\description{
\code{CarefullyCausal} is used to fit outcome regression, inverse probability treatment weighting (IPTW), S-standardization, T-standardization and TMLE.
It prints the causal estimand, causal estimates from the different estimators, causal assumptions and relevant diagnostics.
}
\details{
\code{CarefullyCausal} can be used in a setting with a fixed-exposure (not time-varying). The outcome of interest can
be dichotomous (binomial) or continuous (gaussian) but must be a numeric vector in both settings.
The exposure of interest can be dichotomous, multi-value (maximum 4 levels) or continuous.
When the outcome is dichotomous, the user can specify the argument \code{result_type=""} in order to specify if the output estimates should be in terms of
log-odds (\code{"log"}), odds ratio (\code{"or"}) or risk ratio (\code{"rr"}).
\if{html}{\out{<br>}}
\if{html}{\out{<br>}}
The nature of the exposure variable determines what estimators are deployed.
Currently, TMLE is not yet supported for multi-value exposure and when the exposure is continuous only outcome regression and IPTW are deployed.
}
