% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plot.R
\name{ps.plot}
\alias{ps.plot}
\title{#Propensity score plots}
\usage{
ps.plot(h)
}
\arguments{
\item{h}{Input the saved output object}
}
\value{
Propensity score plots
}
\description{
When exposure is discrete it provides all relevant plots with respect to the reference exposure category
}
